% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{cleanFerries}
\alias{cleanFerries}
\title{Retrieve latest ferry position}
\usage{
cleanFerries(pos_DF)
}
\arguments{
\item{pos_DF}{a data frame created by getPositions}
}
\value{
a data frame with one row per vehicle
}
\description{
Ferries have multiple entries in the live position feed. There is
a note about this on the FAQ. This function filters out the most recent entry
by pulling apart the trip_id column.
}
