% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{fetchNSWRoutes}
\alias{fetchNSWRoutes}
\title{Fetch static gtfs route data from NSW transport.}
\usage{
fetchNSWRoutes()
}
\value{
folder name containing the download
}
\description{
GTFS data comes as a large zip file and can
be used to display all sorts of routes and stops. This function
downloads this data from the NSW transport site. The folder can
then be passed to the functions from the gtfsr package. The gtfsr
package can't be used directly because the transitfeed site doesn't
carry the NSW data anymore. Using the gtfsr functions to process the
results of this requires a modified version - pull request submitted.
}
