% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{getAlerts}
\alias{getAlerts}
\title{extract alerts from a FeedMessage}
\usage{
getAlerts(M)
}
\arguments{
\item{M}{}
}
\value{
dataframe containing alerts
}
\description{
extract alerts from a FeedMessage
}
