% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{getPosition}
\alias{getPosition}
\title{Get a data frame of vehicle positions from a RProtoBuf FeedMessage.}
\usage{
getPosition(M)
}
\arguments{
\item{M}{a FeedMessage, from nswVehicles, or equivalent}
}
\value{
a data frame, possibly with empty columns
}
\description{
Get a data frame of vehicle positions from a RProtoBuf FeedMessage.
}
