% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{getTripUpdates}
\alias{getTripUpdates}
\title{get TripUpdates from a FeedMessage}
\usage{
getTripUpdates(M)
}
\arguments{
\item{M}{}
}
\value{
dataframe containing messages
}
\description{
get TripUpdates from a FeedMessage
}
