% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{nswVehicles}
\alias{nswVehicles}
\title{Vehicle positions for NSW.}
\usage{
nswVehicles(vehicle_type = c("ferries", "sydneytrains"))
}
\value{
RProtoBuf message
}
\description{
This function fetches position data from the NSW public transport
vehicle position API. An API key must be obtained by registering
at \url{https://opendata.transport.nsw.gov.au/}. The result is parsed with the RProtoBuf package using the
Google gtfs-realtime proto. The prototype compiler stuff is initialised at package
load. The result can be used to query trip updates as well as extract positions
}
